# contracts/openapi/auth-service.openapi.yaml
openapi: "3.1.0"
info:
  title: auth-service API
  version: "0.1.0"
  description: Authentication, user management, and RBAC for TaskSphere Pro

paths:
  /auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, maxLength: 255 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "409":
          description: Email already exists

  /auth/login:
    post:
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: Invalid credentials

  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: Invalid or expired refresh token

  /auth/logout:
    post:
      summary: Invalidate refresh token
      security: [{ BearerAuth: [] }]
      responses:
        "204":
          description: Logged out

  /users:
    get:
      summary: List users (Admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
                  total: { type: integer }

  /users/{id}:
    get:
      summary: Get user profile
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      summary: Update user profile or role
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
                role: { type: string, enum: [admin, project_manager, member, guest] }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      summary: Deactivate user (Admin only, soft delete)
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204":
          description: User deactivated

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthTokens:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer }

    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, project_manager, member, guest] }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }